move CopyButton to Successors
