+++ catalog/app/containers/Bucket/PackageTree/PackageTree.tsx
@@ -32,7 +32,6 @@ import { UseQueryResult, useQuery } from 'utils/useQuery'
 import * as workflows from 'utils/workflows'
 
 import Code from '../Code'
-import CopyButton from '../CopyButton'
 import * as Download from '../Download'
 import { FileProperties } from '../FileProperties'
 import * as FileView from '../FileView'
@@ -40,6 +39,7 @@ import Listing, { Item as ListingItem } from '../Listing'
 import PackageCopyDialog from '../PackageCopyDialog'
 import * as PD from '../PackageDialog'
 import Section from '../Section'
+import { CopyButton } from '../Successors'
 import Summary from '../Summary'
 import WithPackagesSupport from '../WithPackagesSupport'
 import * as errors from '../errors'
